Add part attrs to allow applying styles from CSS
At the moment it isn't possible to style the webring HTML from an embedding page's CSS. Adding part="whatever" attributes to the relevant bits does allow that styling from outside, with selectors like `the-claw-webring-widget::part(membersListItem) { background: green; }`, which makes it possible to make the webring fit in with the page. 

If there's some larger plan to have the same styling everywhere and that just hasn't been done yet, then no worries; feel free to reject!